pyproject.toml
[tool.poetry]
name = "metasearch"
version = "0.4.2"
description = "A file management search library"
authors = ["Your Name <your_email@example.com>"]
packages = [
    { include = "metasearch", from = "." }
]
include = ["metasearch/**"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.0.0"
exifread = "^2.3.2"
mutagen = "^1.45.1"
PyMuPDF = "^1.18.19"
python-docx = "^0.8.11"
openpyxl = "^3.0.7"
python-pptx = "^0.6.21"
watchdog = "^2.1.6"
chardet = "^4.0.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
